+++ bin/main.dart
@@ -2,14 +2,17 @@ import 'dart:io';
 
 import 'package:args/args.dart';
 import 'package:screenshots/screenshots.dart';
+import 'package:path/path.dart' as path;
 
-const usage = 'usage: screenshots [--help] [--config <config file>]';
+const usage =
+    'usage: screenshots [-h] [-c <config file>] [-m <normal|recording|comparison|archive>]';
 const sampleUsage = 'sample usage: screenshots';
 
 void main(List<String> arguments) async {
   ArgResults argResults;
 
   final configArg = 'config';
+  final modeArg = 'mode';
   final helpArg = 'help';
   final ArgParser argParser = ArgParser(allowTrailingOptions: false)
     ..addOption(configArg,
@@ -17,12 +20,25 @@ void main(List<String> arguments) async {
         defaultsTo: 'screenshots.yaml',
         help: 'Path to config file.',
         valueHelp: 'screenshots.yaml')
+    ..addOption(modeArg,
+        abbr: 'm',
+        defaultsTo: 'normal',
+        help:
+            'If mode is recording, screenshots will be saved for later comparison. \nIf mode is archive, screenshots will be archived and cannot be uploaded via fastlane.',
+        allowed: ['normal', 'recording', 'comparison', 'archive'],
+        valueHelp: 'normal|recording|comparison|archive')
     ..addFlag(helpArg,
-        help: 'Display this help information.', negatable: false);
+        abbr: 'h', help: 'Display this help information.', negatable: false);
   try {
     argResults = argParser.parse(arguments);
   } on ArgParserException catch (e) {
     _handleError(argParser, e.toString());
+  }
+
+  // show help
+  if (argResults[helpArg]) {
+    _showUsage(argParser);
+    exit(0);
   }
 
   // confirm os
@@ -55,28 +71,15 @@ void main(List<String> arguments) async {
     exit(1);
   }
 
-  // check adb is in path
-  if (!cmd('sh', ['-c', 'which adb && echo adb || echo not installed'], '.',
-          true)
-      .toString()
-      .contains('adb')) {
-    stderr.write(
-        '#############################################################\n');
-    stderr.write("# 'adb' must be in the PATH to use Screenshots\n");
-    stderr.write("# You can usually add it to the PATH using\n"
-        "# export PATH='~/Library/Android/sdk/platform-tools:\$PATH'  \n");
-    stderr.write(
-        '#############################################################\n');
-    exit(1);
-  }
+  // check adb is found
+  getAdbPath();
 
   // validate args
-  final file = File(argResults[configArg]);
-  if (!await file.exists()) {
+  if (!await File(argResults[configArg]).exists()) {
     _handleError(argParser, "File not found: ${argResults[configArg]}");
   }
 
-  await run(argResults[configArg]);
+  await run(argResults[configArg], argResults[modeArg]);
 }
 
 void _handleError(ArgParser argParser, String msg) {
@@ -89,4 +92,28 @@ void _showUsage(ArgParser argParser) {
   print('\n$sampleUsage\n');
   print(argParser.usage);
   exit(2);
+}
+
+/// Path to the `adb` executable.
+String getAdbPath() {
+  final String androidHome = Platform.environment['ANDROID_HOME'] ??
+      Platform.environment['ANDROID_SDK_ROOT'];
+  if (androidHome == null) {
+    throw 'The ANDROID_SDK_ROOT and ANDROID_HOME environment variables are '
+        'missing. At least one of these variables must point to the Android '
+        'SDK directory containing platform-tools.';
+  }
+  final String adbPath = path.join(androidHome, 'platform-tools/adb');
+  final absPath = path.absolute(adbPath);
+  if (!File(adbPath).existsSync()) {
+    stderr.write(
+        '#############################################################\n');
+    stderr.write("# 'adb' must be in the PATH to use Screenshots\n");
+    stderr.write("# You can usually add it to the PATH using\n"
+        "# export PATH='\$HOME/Library/Android/sdk/platform-tools:\$PATH'  \n");
+    stderr.write(
+        '#############################################################\n');
+    exit(1);
+  }
+  return absPath;
 }