+++ lib/src/capture_screen.dart
@@ -1,17 +1,18 @@
 import 'dart:async';
 import 'dart:io';
 
-///
+import 'globals.dart';
+
 /// Called by integration test to capture images.
-///
 Future screenshot(final driver, Map config, String name,
-    {Duration timeout = const Duration(seconds: 30)}) async {
+    {Duration timeout = const Duration(seconds: 30),
+    bool silent = false}) async {
   // todo: auto-naming scheme
-  final stagingDir = config['staging'] + '/test';
   await driver.waitUntilNoTransientCallbacks(timeout: timeout);
-  final List<int> pixels = await driver.screenshot();
-  final File file =
-      await File(stagingDir + '/' + name + '.png').create(recursive: true);
+  final pixels = await driver.screenshot();
+  final testDir = '${config['staging']}/$kTestScreenshotsDir';
+  final file =
+      await File('$testDir/$name.$kImageExtension').create(recursive: true);
   await file.writeAsBytes(pixels);
-  print('Screenshot $name created');
+  if (!silent) print('Screenshot $name created');
 }