+++ lib/src/run.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'archive.dart';
 import 'config.dart';
 import 'daemon_client.dart';
 import 'fastlane.dart' as fastlane;
@@ -20,14 +21,16 @@ import 'utils.dart' as utils;
 /// 3. Process the screenshots including adding a frame if required.
 /// 4. Move processed screenshots to fastlane destination for upload to stores.
 /// 5. If not a real device, stop emulator/simulator.
-Future<void> run([String configPath = kConfigFileName]) async {
+Future<void> run(
+    [String configPath = kConfigFileName, String _runMode = 'normal']) async {
+  final runMode = utils.getRunModeEnum(_runMode);
+
   final screens = Screens();
   await screens.init();
 
   // start flutter daemon
   print('Starting flutter daemon...');
   final daemonClient = DaemonClient();
-  daemonClient.verbose = true;
   await daemonClient.start;
   // get all attached devices and running emulators/simulators
   final devices = await daemonClient.devices;
@@ -44,24 +47,38 @@ Future<void> run([String configPath = kConfigFileName]) async {
 
   // init
   final stagingDir = configInfo['staging'];
-  await Directory(stagingDir + '/test').create(recursive: true);
+  await Directory(stagingDir + '/$kTestScreenshotsDir').create(recursive: true);
   await resources.unpackScripts(stagingDir);
-  await fastlane.clearFastlaneDirs(configInfo, screens);
-  final imageProcessor = ImageProcessor(screens, configInfo);
-
+  final archiveDir = configInfo['archive'];
+  Archive archive = Archive(stagingDir, archiveDir);
+  if (archiveDir == null) {
+    await fastlane.clearFastlaneDirs(configInfo, screens, runMode);
+  } else {
+    print('Archiving screenshots to ${archive.archiveDirPrefix}...');
+  }
   // run integration tests in each real device (or emulator/simulator) for
   // each locale and process screenshots
   await runTestsOnAll(
-      daemonClient, devices, emulators, config, screens, imageProcessor);
+      daemonClient, devices, emulators, config, screens, runMode, archive);
   // shutdown daemon
   await daemonClient.stop;
 
   print('\n\nScreen images are available in:');
-  print('  ios/fastlane/screenshots');
-  print('  android/fastlane/metadata/android');
-  print('for upload to both Apple and Google consoles.');
-  print('\nFor uploading and other automation options see:');
-  print('  https://pub.dartlang.org/packages/fledge');
+  if (runMode == RunMode.recording) {
+    final recordingDir = configInfo['recording'];
+    print('  $recordingDir/ios/fastlane/screenshots');
+    print('  $recordingDir/android/fastlane/metadata/android');
+  } else {
+    if (archiveDir == null) {
+      print('  ios/fastlane/screenshots');
+      print('  android/fastlane/metadata/android');
+      print('for upload to both Apple and Google consoles.');
+      print('\nFor uploading and other automation options see:');
+      print('  https://pub.dartlang.org/packages/fledge');
+    } else {
+      print('  ${archive.archiveDirPrefix}');
+    }
+  }
   print('\nscreenshots completed successfully.');
 }
 
@@ -73,12 +90,35 @@ Future<void> run([String configPath = kConfigFileName]) async {
 /// Assumes the integration tests capture the screen shots into a known directory using
 /// provided [capture_screen.screenshot()].
 Future runTestsOnAll(DaemonClient daemonClient, List devices, List emulators,
-    Config config, Screens screens, ImageProcessor imageProcessor) async {
+    Config config, Screens screens, RunMode runMode, Archive archive) async {
   final configInfo = config.configInfo;
   final locales = configInfo['locales'];
   final stagingDir = configInfo['staging'];
   final testPaths = configInfo['tests'];
   final configDeviceNames = utils.getAllConfiguredDeviceNames(configInfo);
+  final imageProcessor = ImageProcessor(screens, configInfo);
+
+  final recordingDir = configInfo['recording'];
+  final archiveDir = configInfo['archive'];
+  switch (runMode) {
+    case RunMode.normal:
+      break;
+    case RunMode.recording:
+      recordingDir == null
+          ? throw 'Error: \'recording\' dir is not specified in screenshots.yaml'
+          : null;
+      break;
+    case RunMode.comparison:
+      runMode == RunMode.comparison && (!(await utils.isRecorded(recordingDir)))
+          ? throw 'Error: a recording must be run before a comparison'
+          : null;
+      break;
+    case RunMode.archive:
+      archiveDir == null
+          ? throw 'Error: \'archive\' dir is not specified in screenshots.yaml'
+          : null;
+      break;
+  }
 
   for (final configDeviceName in configDeviceNames) {
     // look for matching device first
@@ -87,7 +127,7 @@ Future runTestsOnAll(DaemonClient daemonClient, List devices, List emulators,
     String deviceId;
     Map emulator;
     Map simulator;
-    bool pendingLocaleChange = false;
+    bool pendingIosLocaleChangeAtStart = false;
     if (device != null) {
       deviceId = device['id'];
     } else {
@@ -105,11 +145,11 @@ Future runTestsOnAll(DaemonClient daemonClient, List devices, List emulators,
             utils.getIosSimulators(), configDeviceName);
         deviceId = simulator['udid'];
         // check if current device is pending a locale change
-        if (locales[0] == utils.iosSimulatorLocale(deviceId)) {
+        if (locales[0] == utils.getIosSimulatorLocale(deviceId)) {
           print('Starting $configDeviceName...');
           startSimulator(deviceId);
         } else {
-          pendingLocaleChange = true;
+          pendingIosLocaleChangeAtStart = true;
           print('Not starting $configDeviceName due to pending locale change');
         }
       }
@@ -118,58 +158,53 @@ Future runTestsOnAll(DaemonClient daemonClient, List devices, List emulators,
         ? throw 'Error: device \'$configDeviceName\' not found'
         : null;
 
-    // Check for a running android device or emulator
-    bool isRunningAndroidDeviceOrEmulator(Map device, Map emulator) {
-      return (device != null && device['platform'] != 'ios') ||
-          (device == null && emulator != null);
-    }
-
-    // save original locale for reverting later if necessary
-    String origLocale;
-    if (isRunningAndroidDeviceOrEmulator(device, emulator)) {
-      origLocale = utils.androidDeviceLocale(deviceId);
-    }
-
-    for (final locale in locales) {
-      // set locale if android device or emulator
-      if (isRunningAndroidDeviceOrEmulator(device, emulator)) {
-        await setAndroidLocale(deviceId, locale, configDeviceName);
+    final deviceType = getDeviceType(configInfo, configDeviceName);
+    // if device is real ios or android, cannot change locale
+    if (device != null && !device['emulator']) {
+      final defaultLocale = 'en-US'; // todo: need actual local
+      print('Warning: the locale of a real device cannot be changed.');
+      await runProcessTests(config, screens, configDeviceName, defaultLocale,
+          deviceType, testPaths, deviceId, imageProcessor, runMode, archive);
+    } else {
+      // Check for a running android device or emulator
+      bool isRunningAndroidDeviceOrEmulator(Map device, Map emulator) {
+        return (device != null && device['platform'] != 'ios') ||
+            (device == null && emulator != null);
       }
 
-      // set locale if ios simulator
-      if ((device != null &&
-          device['platform'] == 'ios' &&
-          device['emulator'])) {
-        // an already running simulator
-        await setSimulatorLocale(
-            deviceId, configDeviceName, locale, stagingDir);
-      } else {
-        if (device == null && simulator != null) {
-          if (pendingLocaleChange) {
-            // a non-running simulator
-            await setSimulatorLocale(
-                deviceId, configDeviceName, locale, stagingDir,
-                running: false);
-          } else {
-            // a running simulator
-            await setSimulatorLocale(
-                deviceId, configDeviceName, locale, stagingDir);
+      // save original locale for reverting later if necessary
+      String origAndroidLocale;
+      if (isRunningAndroidDeviceOrEmulator(device, emulator)) {
+        origAndroidLocale = utils.getAndroidDeviceLocale(deviceId);
+      }
+
+      for (final locale in locales) {
+        // set locale if android device or emulator
+        if (isRunningAndroidDeviceOrEmulator(device, emulator)) {
+          await setEmulatorLocale(deviceId, locale, configDeviceName);
+        }
+        // set locale if ios simulator
+        if ((device != null &&
+            device['platform'] == 'ios' &&
+            device['emulator'])) {
+          // an already running simulator
+          await setSimulatorLocale(
+              deviceId, configDeviceName, locale, stagingDir);
+        } else {
+          if (device == null && simulator != null) {
+            if (pendingIosLocaleChangeAtStart) {
+              // a non-running simulator
+              await setSimulatorLocale(
+                  deviceId, configDeviceName, locale, stagingDir,
+                  running: false);
+              pendingIosLocaleChangeAtStart = false;
+            } else {
+              // a running simulator
+              await setSimulatorLocale(
+                  deviceId, configDeviceName, locale, stagingDir);
+            }
           }
         }
-      }
-      // issue locale warning if ios device
-      if ((device != null &&
-          device['platform'] == 'ios' &&
-          !device['emulator'])) {
-        // a running ios device
-        print('Warning: the locale of an ios device cannot be changed.');
-      }
-      final deviceType = getDeviceType(configInfo, configDeviceName);
-
-      // store env for later use by tests
-      // ignore: invalid_use_of_visible_for_testing_member
-      await config.storeEnv(screens, configDeviceName, locale,
-          utils.getStringFromEnum(deviceType));
 
       // Change orientation if required
       final deviceOrientation = configInfo['devices']
@@ -199,26 +234,45 @@ Future runTestsOnAll(DaemonClient daemonClient, List devices, List emulators,
         }
       }
 
-      // run tests
-      for (final testPath in testPaths) {
-        print(
-            'Running $testPath on \'$configDeviceName\' in locale $locale...');
-        await utils.streamCmd('flutter', ['-d', deviceId, 'drive', testPath]);
+      // run tests and process images
+      await runProcessTests(config, screens, configDeviceName, locale,
+          deviceType, testPaths, deviceId, imageProcessor, runMode, archive);
 
-        // process screenshots
-        await imageProcessor.process(deviceType, configDeviceName, locale);
-      }
     }
 
     // if an emulator was started, revert locale if necessary and shut it down
     if (emulator != null) {
-      await setAndroidLocale(deviceId, origLocale, configDeviceName);
+      await setEmulatorLocale(deviceId, origAndroidLocale, configDeviceName);
       await shutdownAndroidEmulator(daemonClient, deviceId);
     }
     if (simulator != null) {
       // todo: revert locale
       shutdownSimulator(deviceId);
     }
+  }
+}
+
+Future runProcessTests(
+    Config config,
+    Screens screens,
+    configDeviceName,
+    String locale,
+    DeviceType deviceType,
+    testPaths,
+    String deviceId,
+    ImageProcessor imageProcessor,
+    RunMode runMode,
+    Archive archive) async {
+  // store env for later use by tests
+  // ignore: invalid_use_of_visible_for_testing_member
+  await config.storeEnv(
+      screens, configDeviceName, locale, utils.getStringFromEnum(deviceType));
+  for (final testPath in testPaths) {
+    print('Running $testPath on \'$configDeviceName\' in locale $locale...');
+    await utils.streamCmd('flutter', ['-d', deviceId, 'drive', testPath]);
+    // process screenshots
+    await imageProcessor.process(
+        deviceType, configDeviceName, locale, runMode, archive);
   }
 }
 
@@ -273,22 +327,22 @@ Future setSimulatorLocale(
     String deviceId, String deviceName, String testLocale, stagingDir,
     {bool running = true}) async {
   // a running simulator
-  final deviceLocale = utils.iosSimulatorLocale(deviceId);
+  final deviceLocale = utils.getIosSimulatorLocale(deviceId);
 //  print('simulator locale=$deviceLocale');
   if (testLocale != deviceLocale) {
     if (running) shutdownSimulator(deviceId);
     print(
         'Changing locale from $deviceLocale to $testLocale on \'$deviceName\'...');
     await _changeSimulatorLocale(stagingDir, deviceId, testLocale);
+    print('Starting $deviceName...');
     startSimulator(deviceId);
   }
 }
 
-/// Set the locale for a real android device or a running emulator.
-Future<void> setAndroidLocale(String deviceId, testLocale, deviceName) async {
-  // a running android device or emulator
-  final deviceLocale = utils.androidDeviceLocale(deviceId);
-//  print('android device or emulator locale=$deviceLocale');
+/// Set the locale of a running emulator.
+Future<void> setEmulatorLocale(String deviceId, testLocale, deviceName) async {
+  final deviceLocale = utils.getAndroidDeviceLocale(deviceId);
+  print('emulator locale=$deviceLocale');
   if (deviceLocale != null &&
       deviceLocale != '' &&
       deviceLocale != testLocale) {
@@ -304,7 +358,7 @@ Future<void> setAndroidLocale(String deviceId, testLocale, deviceName) async {
 /// Change local of real android device or running emulator.
 void changeAndroidLocale(
     String deviceId, String deviceLocale, String testLocale) {
-  if (cmd('adb', ['root'], '.', true) ==
+  if (cmd('adb', ['-s', deviceId, 'root'], '.', true) ==
       'adbd cannot run as root in production builds\n') {
     stdout.write(
         'Warning: locale will not be changed. Running in locale \'$deviceLocale\'.\n');
@@ -313,7 +367,6 @@ void changeAndroidLocale(
     stdout.write(
         '    https://stackoverflow.com/questions/43923996/adb-root-is-not-working-on-emulator/45668555#45668555 for details.\n');
   }
-  _flutterDriverBugWarning();
   // adb shell "setprop persist.sys.locale fr-CA; setprop ctl.restart zygote"
   cmd('adb', [
     '-s',
@@ -332,7 +385,6 @@ void changeAndroidLocale(
 /// Change locale of non-running simulator.
 Future _changeSimulatorLocale(
     String stagingDir, String name, String testLocale) async {
-  _flutterDriverBugWarning();
   await utils.streamCmd('$stagingDir/resources/script/simulator-controller',
       [name, 'locale', testLocale]);
 }
@@ -347,11 +399,6 @@ Future<String> shutdownAndroidEmulator(
     throw 'Error: device id \'$deviceId\' not shutdown';
   }
   return device['id'];
-}
-
-void _flutterDriverBugWarning() {
-  stdout.write(
-      '\nWarning: running tests in a non-default locale will cause test to hang due to a bug in Flutter Driver (not related to \'screenshots\'). Modify your screenshots.yaml to use only the default locale for your location. For details of bug see comment at https://github.com/flutter/flutter/issues/27785#issue-408955077. Give comment a thumbs-up to get it fixed!\n\n');
 }
 
 /// Start android emulator in a CI environment.