+++ test/frame_test.dart
@@ -1,5 +1,5 @@
 import 'package:screenshots/src/config.dart';
-import 'package:screenshots/src/image_magick.dart' as im;
+import 'package:screenshots/src/globals.dart';
 import 'package:screenshots/src/image_processor.dart';
 import 'package:screenshots/src/resources.dart' as resources;
 import 'package:screenshots/src/screens.dart';
@@ -25,7 +25,7 @@ main() {
       'statusbarPath': statusbarPath,
     };
     print('options=$options');
-    await im.imagemagick('overlay', options);
+    await im.convert('overlay', options);
 
     final screenshotNavbarPath =
         '${appConfig['staging']}/${ScreenResources['navbar']}';
@@ -34,7 +34,7 @@ main() {
       'screenshotNavbarPath': screenshotNavbarPath,
     };
     print('options=$options');
-    await im.imagemagick('append', options);
+    await im.convert('append', options);
 
     final framePath = appConfig['staging'] + '/' + ScreenResources['frame'];
     final size = screen['size'];
@@ -49,6 +49,6 @@ main() {
       'backgroundColor': ImageProcessor.kDefaultAndroidBackground,
     };
     print('options=$options');
-    await im.imagemagick('frame', options);
+    await im.convert('frame', options);
   });
 }