+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 1.2.0
+- Added archive feature to collect screenshots of all runs for reporting, etc... #77 #81
+- Improved detection of adb path #79
+
+## 1.1.1
+- Fixed localization issue in test #19, #20
+- Improved handling of locale for emulators and real devices
+
+## 1.1.0
+- Added record/compare feature to compare screenshots with previously recorded screenshots during a run. #65
+
+## 1.0.2
+- Fixed bug with parsing ios simulator info #73
+
 ## 1.0.1
 - Fixed pedantic lint errors
 
@@ -46,7 +60,7 @@
 
 - Bypasses changing locales if running in only one locale
 - Issues warning about running flutter driver in multiple locales  
-  See issue: https://github.com/flutter/flutter/issues/27785 for details.
+  See issue: <https://github.com/flutter/flutter/issues/27785> for details.
 
 ## 0.1.3